requirements.txt
unstructured>=0.10.30
libmagic>=0.4.27
pytesseract>=0.3.10
pdf2image>=1.16.3
python-dotenv>=1.0.0
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-core>=0.1.10
beautifulsoup4>=4.12.2
requests>=2.31.0
python-magic>=0.4.27
slack-sdk>=3.27.1
atlassian-python-api>=3.41.4
slack-bolt>=1.18.0
chromadb>=0.4.22
google-api-python-client>=2.118.0
schedule>=1.2.1
langchain-chroma>=0.0.5
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0